Create Boolean_expression.dart
This program is for the evaluation of the Boolean expression where a string consists of only 0, 1, A, B, C where
A = AND
B = OR
C = XOR
The evaluation is done from left to right. The length of string will be odd. It will always be a valid string.
This pull request is for Hacktober Fest 2020.