Bug fix in factorial calculation
The error is that there is no factorial of a negative number. Therefore, it is necessary to set a test for the sign of the number from which we want to find the factorial.
And one more mistake, the factorial cannot be calculated from real numbers, therefore, in my opinion, instead of var, it is necessary to put in the function